compare, ifCond and for custom Handlebars helpers  were removed from theme.js and moved to custom.handlebars.helpers.js file. added selectIf custom Handlebars helper added comments some formatting
